added service and repository layers to application
