+++ src/main/java/com/codecool/dartcc/repository/PlayerRepository.java
@@ -1,0 +1,4 @@
+package com.codecool.dartcc.repository;
+
+public interface PlayerRepository {
+}
